+++ colossalai/__init__.py
@@ -1,5 +1,5 @@
-from .initialize import (initialize, launch, launch_from_openmpi,
-                         launch_from_slurm, launch_from_torch, get_default_parser)
+from .initialize import (initialize, launch, launch_from_openmpi, launch_from_slurm, launch_from_torch,
+                         get_default_parser, colo_launch, colo_launch_from_torch)
 
+__all__ = ['colo_launch', 'colo_launch_from_torch']
 __version__ = '0.0.1'
-
