+++ colossalai/context/__init__.py
@@ -5,6 +5,6 @@ from .moe_context import MOE_CONTEXT
 from .process_group_initializer import *
 from .random import *
 
-from .distributed_mgr import DISTMGR, colo_launch, colo_launch_from_torch
+from .distributed_mgr import DISTMGR
 
-__all__ = ['DISTMGR', 'colo_launch', 'colo_launch_from_torch']
+__all__ = ['DISTMGR']
